v3.0 Major Refactor/Cleanup
