+++ README.md
@@ -1,7 +1,44 @@
-# FountainCodes over IPv6 Multicast using LT-Codes
+# Sprinkler
 
-__This Repository__ is currently *not under development*. 
+An efficient IPv6 Multicast Protocol for distributing data over a wireless network.
 
-Use [TWIN from ComNets-Bremen](https://github.com/ComNets-Bremen/TWIN) for stable Framework Usage.
+It uses:
 
-> This Code is now know as Sprinkler Protocol for TWIN Nodes for ComNets-Bremen, Bremen, Germany
+1. [__Luby-Transform Codes__](https://en.wikipedia.org/wiki/Luby_transform_code)
+2. [__Trickle Algorithm__](https://tools.ietf.org/html/rfc6206)
+
+## Usability over
+
+* __802.11 WLAN ad-hoc networks with Raspberry Pis and GNU/Linux computers__
+
+## Usage
+
+    python3 Sprinkler.py --version 1 --filename /path/to/file --group ff32::2
+    --port 30002
+
+will distribute the file `/path/to/file` over the wireless ad-hoc network Nodes
+which are running the same code but will lower `version` number (for instance __0__)
+
+### Setup
+
+    sudo ./setup.sh
+
+## Application
+
+* Currently being used in [TWIN Testbed](https://github.com/ComNets-Bremen/TWIN) at the [Sustainable Communication Networks, University of Bremen, Germany](http://comnets.uni-bremen.de) to distribute firmware over __802.11 WLAN back-channel__
+
+## Dependencies
+
+It uses [LT-Code Github Repository by Anson Rosenthal](https://github.com/anrosent/lt-code)
+
+install `pip3` module first using:
+
+    sudo pip3 install lt-code
+
+## References
+
+[Complete Report of the Protocol at ResearchGate](http://dx.doi.org/10.13140/RG.2.2.32561.99687)
+
+### License
+
+published under __GNU Public License v3 (GPLv3)__
