+++ Sprinkler/fountain.py
@@ -1,10 +1,36 @@
 #!/usr/bin/python3
+
+# Sprinkler - Data Dissemination Protocol for Wireless Networks
+# Copyright (C) 2017  Shantanoo Desai
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# This file is a part of Sprinkler
+
+"""
+Fountain Module for Sprinkler
+"""
+
+__author__ = "Shantanoo Desai"
+
 from lt import encode
 from lt.sampler import DEFAULT_DELTA
 from struct import pack
-from math import log, sqrt, floor, ceil
-import TWIN.global_variables as gv
-import datetime, socket
+from math import log, sqrt
+import Sprinkler.global_variables as gv
+import datetime
+import socket
 from os import chdir, path
 import logging
 
@@ -12,11 +38,11 @@ import logging
 logger = logging.getLogger("Fountain")
 logger.setLevel(logging.DEBUG)
 
-handler = logging.FileHandler(path.expanduser("~")+"/logFiles/TWIN.log")
+handler = logging.FileHandler(path.expanduser(".") + "/log/Sprinkler.log")
 handler.setLevel(logging.DEBUG)
 
 # format for logging
-formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+formatter = logging.Formatter('%(name)s - %(levelname)s - %(message)s')
 handler.setFormatter(formatter)
 logger.addHandler(handler)
 
@@ -24,20 +50,21 @@ logger.addHandler(handler)
 def addFooter(encodedBlock, version):
     """
         Function: addFooter
-        params: 
+        params:
         encodedBlock:from the LT-Encoder
         version: 2 Byte value of version to be pack as Footer
-        
+
         return type: Packet of (Blocksize=1452B + 2B_footer)
 
-        Description: 
+        Description:
         Packing a 2 Byte recent version of the Node to
         and LT-encoded Block
     """
 
-    ## Concatenate 2B of Version to the Block
+    # Concatenate 2B of Version to the Block
     packedData = encodedBlock + pack('!H', version)
     return packedData
+
 
 def FounParameters(fname=gv.FILENAME, bsize=gv.BLOCKSIZE):
     """
@@ -56,32 +83,35 @@ def FounParameters(fname=gv.FILENAME, bsize=gv.BLOCKSIZE):
             the file Completely.
     """
 
-    ## Change to Target Path
+    # Change to Target Path
     chdir(gv.PATH)
 
-    ## Open the Target File
+    # Open the Target File
     with open(fname, 'rb') as f:
         fileSize, blockCount = encode._split_file(f, bsize)
 
-    ## Determine Value of K
+    # Determine Value of K
     calculated_K = len(blockCount)
-    logger.debug("FileSize in KB:%0.2f"%(fileSize/1000))
-    logger.debug("No. of Blocks:%d"%calculated_K)
+    logger.debug("FileSize in KB:%0.2f" % (fileSize / 1000))
+    logger.debug("No. of Blocks:%d" % calculated_K)
 
-    ## Determine Value of Gamma
-    calculated_Gamma = (sqrt(calculated_K) * (log(calculated_K/DEFAULT_DELTA))**2/calculated_K)
-    logger.debug("Value of Gamma: %f"%calculated_Gamma)
+    # Determine Value of Gamma
+    calculated_Gamma =\
+        (
+            sqrt(calculated_K) *
+            (log(calculated_K / DEFAULT_DELTA))**2 / calculated_K
+        )
+    logger.debug("Value of Gamma: %f" % calculated_Gamma)
 
     return calculated_K, calculated_Gamma
-
-
 
 
 def CheckConsistency(incomingVersion):
     """
         Function: CheckConsistency
         params:
-        incomingVersion: Incoming 2 Byte version from a neighboring node on the Multicast Channel
+        incomingVersion: Incoming 2 Byte version from a neighboring node
+        on the Multicast Channel
 
         Description:
         Version Check for Consistency based on Trickle Algorithm (RFC6206)
@@ -90,55 +120,53 @@ def CheckConsistency(incomingVersion):
         If We are ahead => Start a Fountain of the Update
     """
 
-
-    logger.debug("theirs: %d, ours:%d"%(incomingVersion, gv.VERSION))
+    logger.debug("theirs: %d, ours:%d" % (incomingVersion, gv.VERSION))
 
     if gv.VERSION == incomingVersion:
-        ## If values are same
-        ## - Consistent
-        if gv.tt.c > gv.tt.k or gv.tt.function.__name__=='fountain':
+        # If values are same
+        # Consistent
+        if gv.tt.c > gv.tt.k or gv.tt.function.__name__ == 'fountain':
 
-            ## If we are in supressed transmission state
-            ## c >= k an if the timer resets
-            ## chances are we might spray an unecessary fountain once
-            ## if that is the case: rather send a TrickleMessage
+            # If we are in supressed transmission state
+            # c >= k an if the timer resets
+            # chances are we might spray an unecessary fountain once
+            # if that is the case: rather send a TrickleMessage
             setattr(gv.tt, 'function', gv.mcastSock.send)
-            setattr(gv.tt, 'kwargs', {'message':pack('!H', gv.VERSION),'host':gv.MCAST_GRP,'port':gv.MCAST_PORT})
+            setattr(gv.tt, 'kwargs', {'message': pack('!H', gv.VERSION),
+                                      'host': gv.MCAST_GRP,
+                                      'port': gv.MCAST_PORT}
+                    )
         logger.info("Consistent")
         gv.tt.hear_consistent()
     else:
 
         if gv.VERSION < incomingVersion:
-            ## If we are Behind
-            ## - Inconsistent message and anticipate
-            ## Update
+            # If we are Behind
+            # - Inconsistent message and anticipate
+            # Update
             logger.info("Lower")
             gv.tt.hear_inconsistent()
 
         else:
-            ## If we are ahead
-            ## - start a Fountain at on the TrickleTimer Instance
-            ## - This is still an Inconsistency
+            # If we are ahead
+            # - start a Fountain at on the TrickleTimer Instance
+            # - This is still an Inconsistency
             logger.info("We are higher, Setup Fountain")
-            
-            """
-            This should be the break-through I am looking for !!
 
-            # What happens now? may change attribute again using tt.function.__name__?
-            """
-            ## set attributes to the already defined Global trickleTimer instance called tt
+            # set attributes to the already defined Global trickleTimer
+            # instance called tt
 
-            ## Set the function
+            # Set the function
             setattr(gv.tt, 'function', fountain)
-            ## Set the Arguments
-            setattr(gv.tt, 'kwargs',{'fname':gv.FILENAME, 'bsize':gv.BLOCKSIZE, 'ver': gv.VERSION})
+            # Set the Arguments
+            setattr(gv.tt, 'kwargs', {'fname': gv.FILENAME,
+                                      'bsize': gv.BLOCKSIZE, 'ver': gv.VERSION}
+                    )
 
             gv.tt.hear_inconsistent()
 
-    ## Pretty much done with the check
+    # Pretty much done with the check
     logger.info("Exiting Consistency Check")
-
-
 
 
 def fountain(fname=gv.FILENAME, bsize=gv.BLOCKSIZE, ver=gv.VERSION):
@@ -160,45 +188,47 @@ def fountain(fname=gv.FILENAME, bsize=gv.BLOCKSIZE, ver=gv.VERSION):
         6. If Limit is reached Stop the Fountain
     """
 
-    ## Step 1:
+    # Step 1:
     k, g = FounParameters(fname, bsize)
 
-    ## Step 2:
+    # Step 2:
     chdir(gv.PATH)
 
     with open(fname, 'rb') as f:
 
-        ## Limit Check Counter
+        # Limit Check Counter
         packetCounter = 0
 
         while True:
-            ## Time Stamp @ beginning
+            # Time Stamp @ beginning
             timeStamp1 = datetime.datetime.now().replace(microsecond=0)
 
             logger.info("Start Fountain")
 
-            ##  Encode Each Block
+            # Encode Each Block
             for eachBlock in encode.encoder(f, bsize):
 
-                ## Step 3:
+                # Step 3:
                 droplet = addFooter(eachBlock, gv.VERSION)
 
                 try:
-                    ## Step 4:
-                    ## Send to all the Multicast Members.
+                    # Step 4:
+                    # Send to all the Multicast Members.
                     gv.mcastSock.send(droplet, gv.MCAST_GRP)
                     packetCounter += 1
 
-                    ## Step 5:
-                    if (packetCounter >= (1+round(g,1))*k):
+                    # Step 5:
+                    if (packetCounter >= (1 + round(g, 1)) * k):
 
-                        ## Time Stamp @ End
-                        timeStamp2 = datetime.datetime.now().replace(microsecond=0)
-                        ## Some stats
-                        logger.debug("Droplets sent %d"%packetCounter)
-                        logger.debug("time needed %s s"%(timeStamp2-timeStamp1))
+                        # Time Stamp @ End
+                        timeStamp2 = datetime.datetime.now().replace(
+                            microsecond=0)
+                        # Some stats
+                        logger.debug("Droplets sent %d" % packetCounter)
+                        logger.debug("time needed %s s"
+                                     % (timeStamp2 - timeStamp1))
 
-                        ## Step 6:
+                        # Step 6:
                         logger.info("Closing Fountain")
                         break
