+++ Sprinkler/trickle.py
@@ -19,6 +19,11 @@
 # not limited to the correctness, accuracy, reliability or usefulness of
 # this software.
 
+__author__ = "Tony Cheaneau"
+__copyright__ = "None"
+__license__ = "Public Domain Contribution"
+__email__ = "tony.cheneau@nist.gov"
+
 """(Somewhat generic) Trickle timer (See RFC 6206)"""
 from random import uniform
 from threading import Timer, RLock
@@ -55,7 +60,8 @@ class trickleTimer(object):
         # store the function, so that it can be rescheduled multiple times
         self.function = function
         self.kwargs = kwargs
-        logger.debug("next trickle timer is set to run in t = %f seconds" % self.t)
+        logger.debug("next trickle timer is set to run in t = %f seconds"
+                     % self.t)
         self.lock = RLock()
         self.thread = Timer(self.t, self.__run)
         self.thread.daemon = True
@@ -96,7 +102,8 @@ class trickleTimer(object):
         """Trickle timer has expired"""
         with self.lock:
             # step 5
-            logger.debug("trickle timer has expired, increasing minimum interval size")
+            logger.debug("trickle timer has expired, increasing minimum\
+                         interval size")
             self.I = self.I * 2
             if self.I > self.Imax:
                 logger.info("trickle timer has reached maximum interval size")
@@ -107,7 +114,8 @@ class trickleTimer(object):
 
             # set the new timer
             self.t = uniform(self.I / 2, self.I)
-            logging.debug("next trickle timer is set to run in %f seconds" % self.t)
+            logging.debug("next trickle timer is set to run in %f seconds"
+                          % self.t)
 
             try:
                 self.thread.cancel()  # should never be needed
