--- TWIN.py
@@ -1,6 +1,0 @@
-import sys
-
-from TWIN.main import main
-
-if __name__ == '__main__':
-    main(sys.argv)
